Update savedmodel_test.py
I added a cast to `float32`. This is needed because of an obscure bug in JAX (#6874).

+++ jax/experimental/jax2tf/tests/savedmodel_test.py
@@ -83,7 +83,7 @@ class SavedModelTest(tf_test_util.JaxToTfTestCase):
       x, = primals
       x_dot, = tangents
       primal_out = f_jax(x)
-      tangent_out = 3. * x * x_dot
+      tangent_out = np.float32(3.) * x * x_dot
       return primal_out, tangent_out
 
     model = tf.Module()
